# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 电商客服 Agent API 契约
  version: 0.1.0
  summary: 第一版外部决策、补上下文、人工反馈、客户 Admin 和系统 Admin 核心接口。
  description: |
    本契约从 `docs/http-api-design.md`、`docs/customer-admin-design.md` 和 `docs/system-admin-design.md` 抽出，覆盖第一版主链路、客户 Admin、系统 Admin 和商品资料核心接口。
    外部系统通过 `/v1/reply-decisions` 创建决策，按 `context_requests[]` 回填商品、订单、物流、规则上下文，并通过 `actions/results` 回传外部动作结果。
    `decision_id` 是外部 API 主键，同时映射内部 LangGraph `thread_id`；`graph_version` 必须写入 trace 和 checkpoint，便于旧决策回放。

    鉴权域必须隔离：外部系统 API token 只能调用外部决策接口，客户 Admin session 只能调用 `/v1/admin/*` 和客户可见追踪接口，系统 Admin session 只能调用 `/v1/system-admin/*`。三类凭据不能混用。
    客户 Admin Web 与系统 Admin Web 必须使用不同站点、登录页、Cookie / session 名和路由守卫；客户后台 session 建议命名为 `agent_admin_session`，系统后台 session 建议命名为 `agent_system_admin_session`。
servers:
  - url: https://api.example.com
    description: 生产或 dev 环境 API 域名占位，实际部署以环境配置为准。
...
paths:
...
  /v1/system-admin/llm/providers:
    get:
      tags: [System Admin LLM Governance]
      operationId: listSystemAdminLlmProviders
      summary: 查询 Provider 配置
      description: super_admin、release_admin、technical_support 和 security_auditor 可读；仅返回 Secret 引用，不返回密钥内容。
      x-roles: [super_admin, release_admin, technical_support, security_auditor]
      security:
        - SystemAdminSession: []
      responses:
        "200":
          description: Provider 列表；无配置时 items 为空数组。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProviderListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [System Admin LLM Governance]
      operationId: createSystemAdminLlmProvider
      summary: 创建 Provider 配置
      description: 仅 super_admin 或 release_admin 可创建；Secret 只能以 Kubernetes namespace、name、key 引用。
      x-roles: [super_admin, release_admin]
      security:
        - SystemAdminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmProviderCreateRequest"
            example:
              name: primary
              provider_type: openai_compatible
              base_url: https://provider.example/v1
              secret_ref: {namespace: runtime, name: llm-provider, key: api-key}
              enabled: true
              reason: configure primary provider
              idempotency_key: provider-create-001
      responses:
        "201":
          description: Provider 已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/providers/{provider_id}:
    patch:
      tags: [System Admin LLM Governance]
      operationId: updateSystemAdminLlmProvider
      summary: 更新 Provider 名称或启用状态
      description: 仅 super_admin 或 release_admin 可更新；expected_revision 防止覆盖并发修改，端点和 Secret 引用不可原地替换。
      x-roles: [super_admin, release_admin]
      security:
        - SystemAdminSession: []
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmProviderUpdateRequest"
            example:
              expected_revision: 3
              enabled: false
              reason: temporarily disable an unhealthy provider
              idempotency_key: provider-update-001
      responses:
        "200":
          description: Provider 已更新。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/providers/{provider_id}/connection-tests:
    post:
      tags: [System Admin LLM Governance]
      operationId: testSystemAdminLlmProviderConnection
      summary: 对草稿执行 Provider 连接测试
      description: super_admin、release_admin 或 technical_support 可执行。config_version_id 必须指向草稿；只记录状态、耗时和脱敏错误码，不记录请求或响应正文。
      x-roles: [super_admin, release_admin, technical_support]
      security:
        - SystemAdminSession: []
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmConnectionTestRequest"
            example:
              config_version_id: 22222222-2222-4222-8222-222222222222
              timeout_seconds: 20
              max_tokens: 128
              reason: verify the draft provider reference
              idempotency_key: connection-test-001
      responses:
        "202":
          description: 连接测试已完成并记录脱敏结果。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConnectionTest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/config-versions:
    get:
      tags: [System Admin LLM Governance]
      operationId: listSystemAdminLlmConfigVersions
      summary: 按组织查询 LLM 配置版本
      description: 返回不可变版本历史及其 release_record 和 evaluation 摘要，不伪造不存在的发布记录。
      x-roles: [super_admin, release_admin, technical_support, security_auditor]
      security:
        - SystemAdminSession: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema: {type: string, format: uuid, minLength: 36, maxLength: 36}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: cursor
          in: query
          description: Opaque cursor for the exclusive page boundary. Reuse only with the same organization_id; changing the organization or supplying an invalid or mismatched scope returns 422.
          schema: {type: string, minLength: 1, maxLength: 1024}
      responses:
        "200":
          description: 版本列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConfigVersionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/releases:
    get:
      tags: [System Admin LLM Governance]
      operationId: listSystemAdminLlmReleaseRecords
      summary: 按组织查询真实 LLM 发布记录
      description: super_admin、release_admin、technical_support 和 security_auditor 可读；返回 llm_release_record 持久化记录，不从配置版本状态推断。
      x-roles: [super_admin, release_admin, technical_support, security_auditor]
      security:
        - SystemAdminSession: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema: {type: string, format: uuid, minLength: 36, maxLength: 36}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: cursor
          in: query
          description: Opaque exclusive cursor；只能在相同 organization_id 下复用，无下一页时为 null。
          schema: {type: string, minLength: 1, maxLength: 1024}
      responses:
        "200":
          description: 真实发布记录 cursor 分页列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmReleaseRecordListResponse"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationError"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/system-admin/llm/config-versions/drafts:
    post:
      tags: [System Admin LLM Governance]
      operationId: createSystemAdminLlmConfigDraft
      summary: 创建组织级配置草稿
      description: super_admin 或 release_admin 创建新草稿；所有写操作携带审计原因和幂等键。
      x-roles: [super_admin, release_admin]
      security:
        - SystemAdminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmConfigDraftCreateRequest"
            example:
              organization_id: 11111111-1111-4111-8111-111111111111
              description: customer reply model tuning
              reason: prepare a reviewed configuration change
              idempotency_key: config-draft-001
      responses:
        "201":
          description: 草稿已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConfigVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/config-versions/{version_id}:
    get:
      tags: [System Admin LLM Governance]
      operationId: getSystemAdminLlmConfigVersion
      summary: 查询单个配置版本
      description: 返回路由、运行参数、发布记录和评测引用；不返回 Prompt、消息正文或密钥内容。
      x-roles: [super_admin, release_admin, technical_support, security_auditor]
      security:
        - SystemAdminSession: []
      parameters:
        - $ref: "#/components/parameters/LlmVersionId"
      responses:
        "200":
          description: 配置版本。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConfigVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/config-versions/{version_id}/routes:
    put:
      tags: [System Admin LLM Governance]
      operationId: replaceSystemAdminLlmRoutes
      summary: 完整替换草稿场景路由
      description: 仅草稿可改；expected_revision 防止并发覆盖，必须提供各场景主模型、可选降级模型及完整运行参数。
      x-roles: [super_admin, release_admin]
      security:
        - SystemAdminSession: []
      parameters:
        - $ref: "#/components/parameters/LlmVersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmRoutesReplaceRequest"
            example:
              expected_revision: 1
              routes:
                - scenario: reply_generation
                  primary_provider_config_id: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
                  primary_model: chat-pro
                  fallback_provider_config_id: null
                  fallback_model: null
                  enabled: true
                  temperature: 0.2
                  max_output_tokens: 1200
                  timeout_seconds: 18
                  max_retries: 2
                  circuit_breaker_threshold: 5
                  recovery_probe_seconds: 30
              reason: configure reviewed scenario routes
              idempotency_key: config-routes-put-001
      responses:
        "200":
          description: 路由已替换。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConfigVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [System Admin LLM Governance]
      operationId: patchSystemAdminLlmRoutes
      summary: 以完整集合更新草稿场景路由
      description: 与 PUT 使用同一完整替换语义；不支持对已验证或已发布版本原地修改。
      x-roles: [super_admin, release_admin]
      security:
        - SystemAdminSession: []
      parameters:
        - $ref: "#/components/parameters/LlmVersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmRoutesReplaceRequest"
            example:
              expected_revision: 2
              routes:
                - scenario: knowledge_extraction
                  primary_provider_config_id: aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa
                  primary_model: extract-pro
                  fallback_provider_config_id: bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb
                  fallback_model: extract-lite
                  enabled: true
                  temperature: 0.1
                  max_output_tokens: 800
                  timeout_seconds: 20
                  max_retries: 1
                  circuit_breaker_threshold: 5
                  recovery_probe_seconds: 30
              reason: update the complete draft route set
              idempotency_key: config-routes-patch-001
      responses:
        "200":
          description: 路由已替换。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConfigVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/config-versions/{version_id}/validate:
    post:
      tags: [System Admin LLM Governance]
      operationId: validateSystemAdminLlmConfig
      summary: 验证配置草稿
      description: 执行 draft 到 validated；要求场景完整且每个 Provider 对当前草稿有通过的连接测试。
      x-roles: [super_admin, release_admin]
      security:
        - SystemAdminSession: []
      parameters:
        - $ref: "#/components/parameters/LlmVersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmRevisionWriteRequest"
            example:
              expected_revision: 2
              reason: all required routes and connection checks passed
              idempotency_key: config-validate-001
      responses:
        "200":
          description: 草稿已验证。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConfigVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/config-versions/{version_id}/submit-publish:
    post:
      tags: [System Admin LLM Governance]
      operationId: submitSystemAdminLlmConfigPublish
      summary: 提交已验证配置发布
      description: 执行 validated 到 pending_publish；只有真实 evaluation_run_id 通过 release gate 后才创建 release_record。
      x-roles: [super_admin, release_admin]
      security:
        - SystemAdminSession: []
      parameters:
        - $ref: "#/components/parameters/LlmVersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmSubmitPublishRequest"
            example:
              expected_revision: 3
              evaluation_run_id: eval-release-20260714
              reason: evaluation gate passed
              idempotency_key: config-submit-001
      responses:
        "200":
          description: 配置已进入待发布。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConfigVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/system-admin/llm/config-versions/{version_id}/publish:
    post:
      tags: [System Admin LLM Governance]
      operationId: publishSystemAdminLlmConfig
      summary: 发布待发布配置
      description: 仅 super_admin 或 release_admin 可执行 pending_publish 到 running；失败时原运行版本不切换，expected_revision 防止并发覆盖。
      x-roles: [super_admin, release_admin]
      security:
        - SystemAdminSession: []
      parameters:
        - $ref: "#/components/parameters/LlmVersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LlmRevisionWriteRequest"
            example:
              expected_revision: 4
              reason: release approved for production traffic
              idempotency_key: config-publish-001
      responses:
        "200":
          description: 配置已成为运行版本。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LlmConfigVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    AgentApiAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key or Connector Token
      description: 外部系统接入凭据，只能调用外部决策、补上下文、动作结果和反馈接口，不能调用客户 Admin 或系统 Admin API。Connector Token 调用 `POST /v1/reply-decisions` 时必须带有效 `billing_lease`。
    OpenErpIntegrationAuth:
      type: http
      scheme: bearer
      bearerFormat: Service Token
      description: open_erp_agent 服务间 provisioning 凭据，只能调用 `/v1/integrations/open-erp/*`，不能调用 Admin API，也不能替代 Connector Token 调用决策接口。
    AdminSession:
      type: apiKey
      in: cookie
      name: agent_admin_session
      description: 客户 Admin HttpOnly session cookie。session 状态必须落外部持久化存储，不能依赖单容器内存。
    SystemAdminSession:
      type: apiKey
      in: cookie
      name: agent_system_admin_session
      description: 系统 Admin HttpOnly session cookie。不能用于客户 Admin 或外部系统 API。
...
  responses:
...
    Unauthorized:
      description: 未登录、凭据缺失或凭据无效。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 已认证但没有访问当前租户、店铺、角色或排障数据的权限。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在或不属于当前租户 / 店铺上下文。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: 资源状态冲突、唯一约束冲突或业务版本冲突。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: 字段校验失败。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: 服务端内部错误。响应不得包含密钥、请求头、Cookie 或完整 raw payload。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    LlmProviderListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: "#/components/schemas/LlmProvider"}
...